+++ wrappers/Python/roadrunner/PyEventListener.h
@@ -80,13 +80,13 @@ public:
 		}
     }
 
-    void fireOnTrigger(ExecutableModel* model, int index, const std::string& eventId)
+    void fireOnTrigger(ExecutableModel* model, size_t index, const std::string& eventId)
     {
         Log(Logger::LOG_INFORMATION) << __FUNC__;
         onTrigger(model, index, eventId);
     }
 
-    void fireOnAssignment(ExecutableModel* model, int index, const std::string& eventId)
+    void fireOnAssignment(ExecutableModel* model, size_t index, const std::string& eventId)
     {
         Log(Logger::LOG_INFORMATION) << __FUNC__;
         onAssignment(model, index, eventId);
@@ -97,7 +97,7 @@ private:
     PyObject *pyOnTrigger;
     PyObject *pyOnAssignment;
 
-    virtual uint onTrigger(ExecutableModel* model, int index, const std::string& eventId)
+    virtual uint onTrigger(ExecutableModel* model, size_t index, const std::string& eventId)
     {
         uint result = 0;
         std::string err;
@@ -146,7 +146,7 @@ private:
             Py_XDECREF(pyres);
             Py_XDECREF(args);
 
-            // Release the thread. No Python API allowed beyond this point.
+            // Release the thread. No Python wrappers allowed beyond this point.
             PyGILState_Release(gstate);
 
             if (!err.empty())
@@ -158,7 +158,7 @@ private:
         return result;
     }
 
-    virtual uint onAssignment(ExecutableModel* model, int index, const std::string& eventId)
+    virtual uint onAssignment(ExecutableModel* model, size_t index, const std::string& eventId)
     {
         uint result = 0;
         std::string err;
@@ -199,7 +199,7 @@ private:
             Py_XDECREF(pyres);
             Py_XDECREF(args);
 
-            // Release the thread. No Python API allowed beyond this point.
+            // Release the thread. No Python wrappers allowed beyond this point.
             PyGILState_Release(gstate);
 
             if (!err.empty())
